Conflicts:
	letsgo-client/src/main/java/org/igo/client/Main.java
	letsgo-ejb/src/test/java/org/igo/ejb/GoGameTest.java
	letsgo-jpa/src/main/java/org/igo/jpa/GoGame.java

+++ letsgo-remote/src/main/java/org/igo/remote/intf/IGameRemote.java
@@ -16,6 +16,6 @@ public interface IGameRemote {
 
     String getString();
 
-    Integer create();
+    Long create();
 
 }
